backend/requirements.txt
# ResumeCraft Backend Requirements
# Python 3.10+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6

# LangChain & LLMs
langchain==0.1.9
langchain-openai==0.0.8
langchain-core>=0.1.26,<0.2
langchain-community==0.0.24
langgraph==0.0.28

# OpenAI
openai==1.12.0

# Environment & Config
python-dotenv==1.0.1

# Logging & Monitoring
loguru==0.7.2

# HTTP & Networking
httpx==0.26.0
aiohttp==3.9.3

# Data Processing
pydantic-core==2.16.2

# Database (Optional - for future persistence)
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Store (Optional - for future semantic search)
chromadb==0.4.22
sentence-transformers==2.3.1

# File Processing (Optional - for PDF/DOCX parsing)
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Utilities
tiktoken==0.6.0
tenacity==8.2.3

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Production
gunicorn==21.2.0
